ナビゲーションバーを別ファイルに
