+++ lib/main.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:soccer_app_flutter/pages/menu_page.dart';
-import 'package:soccer_app_flutter/pages/connection_page.dart';
-import 'package:soccer_app_flutter/pages/note_page.dart';
+import 'package:soccer_app_flutter/pages/main_navigation_page.dart';
 import 'package:soccer_app_flutter/theme/button_theme_extension.dart';
 
 void main() {
@@ -29,50 +27,6 @@ class MyApp extends StatelessWidget {
         ],
       ),
       home: const MainNavigationPage(title: 'Connection Demo Page'),
-    );
-  }
-}
-
-class MainNavigationPage extends StatefulWidget {
-  const MainNavigationPage({super.key, required this.title});
-  final String title;
-
-  @override
-  State<MainNavigationPage> createState() => MainNavigationPageState();
-}
-
-class MainNavigationPageState extends State<MainNavigationPage> {
-  int _selectedIndex = 1; // デフォルは接続画面（インデックス1）
-
-  // 各画面のリスト
-  final List<Widget> _pages = [
-    const MenuPage(),
-    const ConnectionPage(),
-    const NotePage(),
-  ];
-
-  void onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index; // 選択されたインデックスを更新
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: _pages[_selectedIndex], // 選択されたページを表示
-      bottomNavigationBar: BottomNavigationBar(
-        items: const <BottomNavigationBarItem>[
-          BottomNavigationBarItem(icon: Icon(Icons.menu), label: 'メニュー'),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.bluetooth_connected),
-            label: '接続',
-          ),
-          BottomNavigationBarItem(icon: Icon(Icons.settings), label: '自由帳'),
-        ],
-        currentIndex: _selectedIndex, // 現在のインデックスを設定
-        onTap: onItemTapped, // アイテムがタップされたときの処理
-      ),
     );
   }
 }
